requirements.txt
# ── Core runtime ──
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ── Media processing ──
ffmpeg-python>=0.2.0
opencv-python>=4.9.0
Pillow>=10.2.0
scenedetect[opencv]>=0.6.2

# ── Audio analysis ──
librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.26.0

# ── ML / Clip ranking ──
torch>=2.1.0
torchvision>=0.16.0
open-clip-torch>=2.24.0

# ── Downloading ──
yt-dlp>=2024.1.0

# ── API / Web ──
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
duckduckgo-search>=6.2.0

# ── Database ──
aiosqlite>=0.19.0

# ── Utilities ──
rich>=13.7.0
click>=8.1.0
tqdm>=4.66.0
